finished HW7, my s-prints were the issue....

+++ hdr/vector.h
@@ -22,8 +22,8 @@ typedef enum {
 typedef struct {
   vector_t type;
   void **data;
-  size_t size;
-  size_t capacity;
+  int size;
+  int capacity;
 } vector;
 
 void vector_init(vector *v, size_t capacity, vector_t type);
